refactor(services): move interfaces to a folder in the services folder
