+++ src/services/interfaces/IApplicationsService.ts
