+++ src/services/interfaces/IClipboardHistoryService.ts
