+++ src/services/interfaces/IExtensionManager.ts
