+++ src/services/interfaces/ILogService.ts
