+++ src/services/interfaces/IWindowService.ts
